refactor(sanity): move `useVersionRelease` to shared location
