+++ packages/sanity/src/core/hooks/index.ts
@@ -29,4 +29,5 @@ export * from './useTools'
 export * from './useUnitFormatter'
 export * from './useUserListWithPermissions'
 export * from './useValidationStatus'
+export * from './useVersionRelease'
 export * from './useWorkspaceSchemaId'
